+++ app/components/app/character/Username.vue
@@ -8,11 +8,13 @@ const copy = useCopyUsername(props.username)
 </script>
 
 <template>
-  <button
-    v-if="showCopy && username" size="sm" class="text-blue-500 hover:text-blue-600 flex items-center gap-1"
-    @click.stop.prevent="copy"
+  <component
+    :is="showCopy ? 'button' : 'span'"
+    size="sm"
+    class="text-blue-500 hover:text-blue-600 flex items-center gap-1"
+    @click.stop.prevent="showCopy && copy"
   >
     @{{ username }}
-    <Icon name="content_copy" style="font-size: 1.15rem" />
-  </button>
+    <Icon v-if="showCopy" name="content_copy" style="font-size: 1.15rem" />
+  </component>
 </template>