+++ app/components/ui/data-input/Textarea.vue
@@ -19,10 +19,22 @@ const props = defineProps<Props>()
 const label = useSlotContent(() => props.label)
 
 const { value, errorMessage, handleBlur } = useField(props.path, props.rules)
+
+const labelRef = ref<HTMLInputElement>()
+
+function focus() {
+  if (labelRef.value) {
+    labelRef.value?.focus()
+  }
+}
+
+defineExpose({
+  focus,
+})
 </script>
 
 <template>
-  <label class="form-control w-full">
+  <label ref="labelRef" class="form-control w-full">
     <div v-if="label" class="label">
       <span class="label-text">
         {{ label }}